RF | POM adding BaseProjectPage
