+++ src/test/java/school/redrover/model/OrganizationFolderDetailsPage.java
@@ -5,44 +5,12 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import school.redrover.model.base.BasePage;
+import school.redrover.model.base.BaseProjectPage;
 
-public class OrganizationFolderDetailsPage extends BasePage {
-
-    @FindBy(xpath = "//h1")
-    private WebElement projectName;
-
-    @FindBy(xpath = "//a[contains(@href, '/confirm-rename')]")
-    private WebElement renameSubmenu;
-
-    @FindBy(name = "Submit")
-    private WebElement disableButton;
-
-    @FindBy(id = "enable-project")
-    private WebElement disableMessage;
+public class OrganizationFolderDetailsPage extends BaseProjectPage {
 
     public OrganizationFolderDetailsPage(WebDriver driver) {
         super(driver);
     }
 
-    public String getProjectName() {
-
-        return projectName.getText();
-    }
-
-    public OrganizationFolderRenamePage clickRenameOptionFromLeftSideMenu() {
-        renameSubmenu.click();
-
-        return new OrganizationFolderRenamePage(getDriver());
-    }
-
-    public OrganizationFolderDetailsPage clickDisableButton() {
-        disableButton.click();
-
-        return this;
-    }
-
-    public String getDisabledMessageText() {
-
-        return disableMessage.getText().substring(0, 46);
-    }
 }
